requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.8
attrs==25.1.0
autopep8==2.3.2
bandit==1.8.2
bcrypt==4.2.0
beanie==1.29.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dill==0.3.9
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cli==0.0.7
filelock==3.16.1
frozenlist==1.5.0
git-filter-repo==2.38.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.4
httpx==0.25.0
identify==2.6.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.5
jwt==1.3.1
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mongoengine==0.29.1
motor==3.7.0
multidict==6.1.0
mutagen==1.47.0
mysql-connector-python==9.1.0
ndg-httpsclient==0.5.1
nodeenv==1.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pbr==6.1.0
Pillow==10.1.0
platformdirs==4.3.6
pluggy==1.5.0
polars==1.13.1
pre_commit==4.0.1
propcache==0.2.1
pusher==3.3.2
pyasn1==0.4.8
pycodestyle==2.12.1
pycparser==2.22
pycryptodomex==3.19.0
pydantic==2.9.2
pydantic-settings==2.1.0
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.4.0
pymongo==4.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyonfx==0.9.13
pyOpenSSL==23.3.0
pyquaternion==0.9.9
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.17
pytz==2024.2
pywin32==306
PyYAML==6.0.2
requests==2.31.0
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
ruff==0.7.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
stevedore==5.4.0
toml==0.10.2
tomlkit==0.13.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.1.0
uvicorn==0.32.0
virtualenv==20.27.1
watchfiles==1.0.3
websockets==12.0
winput==1.5.0
yarl==1.18.3
yt-dlp==2023.11.16
